rename catalog-search-connector to product-page-search-connector
