--- bundles/catalog-search-connector/README.md
@@ -1,12 +1,0 @@
-# catalog-search-connector
-[![license](https://img.shields.io/github/license/fond-of-kudu/catalog-search-connector.svg)](https://packagist.org/packages/fond-of-kudu/catalog-search-connector)
-
-## What it does
-
-contains various plugins for extending and searching within Elasticsearch.
-
-## Installation
-
-```
-composer require fond-of-kudu/catalog-search-connector
-```
