--- bundles/catalog-search-connector/src/FondOfKudu/Client/CatalogSearchConnector/Plugin/Elasticsearch/Sorting/AscendingAvailableSortConfigTransferBuilderPlugin.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace FondOfKudu\Client\CatalogSearchConnector\Plugin\Elasticsearch\Sorting;
-
-use Generated\Shared\Search\PageIndexMap;
-use Generated\Shared\Transfer\SortConfigTransfer;
-use Spryker\Client\Catalog\Dependency\Plugin\SortConfigTransferBuilderPluginInterface;
-use Spryker\Client\Kernel\AbstractPlugin;
-
-class AscendingAvailableSortConfigTransferBuilderPlugin extends AbstractPlugin implements SortConfigTransferBuilderPluginInterface
-{
-    /**
-     * @var string
-     */
-    public const PARAMETER_NAME = 'available';
-
-    /**
-     * @return \Generated\Shared\Transfer\SortConfigTransfer
-     */
-    public function build(): SortConfigTransfer
-    {
-        return (new SortConfigTransfer())
-            ->setName('available')
-            ->setParameterName(static::PARAMETER_NAME)
-            ->setFieldName(PageIndexMap::INTEGER_SORT)
-            ->setIsDescending(false);
-    }
-}
