+++ bundles/product-page-search-connector/.gitignore
