+++ bundles/product-page-search-connector/README.md
@@ -1,0 +1,12 @@
+# product-page-search-connector
+[![license](https://img.shields.io/github/license/fond-of-kudu/product-page-search-connector.svg)](https://packagist.org/packages/fond-of-kudu/product-page-search-connector)
+
+## What it does
+
+contains various plugins for extending and searching within Elasticsearch.
+
+## Installation
+
+```
+composer require fond-of-kudu/product-page-search-connector
+```
