+++ bundles/product-page-search-connector/codeception.yml
@@ -1,0 +1,43 @@
+namespace: fond-of-kudu/product-page-search-connector
+
+suites:
+  unit:
+    path: .
+
+settings:
+  shuffle: true
+  lint: true
+
+bootstrap: _bootstrap.php
+
+paths:
+  tests: tests
+  output: tests/_output
+  support: tests/_support
+  data: tests/_data
+
+coverage:
+  enabled: true
+  include:
+    - src/*.php
+
+modules:
+  enabled:
+    - \FondOfCodeception\Module\Spryker
+  config:
+    \FondOfCodeception\Module\Spryker:
+      generate_transfer: false
+      generate_map_classes: false
+      generate_propel_classes: false
+      generate_ide_auto_completion: false
+
+env:
+  standalone:
+    modules:
+      config:
+        \FondOfCodeception\Module\Spryker:
+          generate_transfer: true
+          generate_map_classes: true
+          generate_propel_classes: true
+          generate_ide_auto_completion: true
+  monorepo:
