+++ bundles/product-page-search-connector/phpcs.xml
