+++ bundles/product-page-search-connector/src/FondOfKudu/Shared/ProductPageSearchConnector/ProductPageSearchConnectorConstants.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace FondOfKudu\Shared\ProductPageSearchConnector;
+
+interface ProductPageSearchConnectorConstants
+{
+    /**
+     * @var string
+     */
+    public const ATTR_AVAILABLE = 'available';
+
+    /**
+     * @var string
+     */
+    public const ATTR_IS_SOLD_OUT = 'is_sold_out';
+
+    /**
+     * @var string
+     */
+    public const PRODUCT_DATA_LOCALE = 'locale';
+
+    /**
+     * @var string
+     */
+    public const PRODUCT_DATA_ID_PRODUCT_ABSTRACT = 'id_product_abstract';
+
+    /**
+     * @var string
+     */
+    public const PRODUCT_DATA_STORE = 'store';
+
+    /**
+     * @var string
+     */
+    public const PRODUCT_DATA_SKU = 'sku';
+}
