+++ bundles/product-page-search-connector/src/FondOfKudu/Shared/ProductPageSearchConnector/Schema/page.json
