+++ bundles/product-page-search-connector/src/FondOfKudu/Shared/ProductPageSearchConnector/Transfer/product_page_search_connector.transfer.xml
