+++ bundles/product-page-search-connector/src/FondOfKudu/Zed/ProductPageSearchConnector/Dependency/Facade/ProductPageSearchConnectorToAvailabilityFacadeInterface.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace FondOfKudu\Zed\ProductPageSearchConnector\Dependency\Facade;
+
+use Generated\Shared\Transfer\ProductAbstractAvailabilityTransfer;
+use Generated\Shared\Transfer\StoreTransfer;
+
+interface ProductPageSearchConnectorToAvailabilityFacadeInterface
+{
+    /**
+     * @param string $sku
+     * @param \Generated\Shared\Transfer\StoreTransfer $storeTransfer
+     *
+     * @return \Generated\Shared\Transfer\ProductAbstractAvailabilityTransfer|null
+     */
+    public function findOrCreateProductAbstractAvailabilityBySkuForStore(
+        string $sku,
+        StoreTransfer $storeTransfer
+    ): ?ProductAbstractAvailabilityTransfer;
+}
