+++ bundles/product-page-search-connector/src/FondOfKudu/Zed/ProductPageSearchConnector/Dependency/Facade/ProductPageSearchConnectorToStoreFacadeInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace FondOfKudu\Zed\ProductPageSearchConnector\Dependency\Facade;
+
+use Generated\Shared\Transfer\StoreTransfer;
+
+interface ProductPageSearchConnectorToStoreFacadeInterface
+{
+    /**
+     * @param string $storeName
+     *
+     * @return \Generated\Shared\Transfer\StoreTransfer|null
+     */
+    public function findStoreByName(string $storeName): ?StoreTransfer;
+}
