+++ composer.json
@@ -61,7 +61,7 @@
   },
   "replace": {
     "fond-of-kudu/carts-rest-api": "*",
-    "fond-of-kudu/catalog-search-connector": "*",
+    "fond-of-kudu/product-page-search-connector": "*",
     "fond-of-kudu/checkout-rest-api": "*",
     "fond-of-kudu/checkout-rest-api-country-connector": "*",
     "fond-of-kudu/checkout-data-gift-cart-payment-country-filter": "*",
@@ -92,9 +92,9 @@
       "FondOfKudu\\Zed\\CartsRestApi\\": "bundles/carts-rest-api/src/FondOfKudu/Zed/CartsRestApi",
       "FondOfKudu\\Shared\\CartsRestApi\\": "bundles/carts-rest-api/src/FondOfKudu/Shared/CartsRestApi",
       "FondOfKudu\\Client\\CartsRestApi\\": "bundles/carts-rest-api/src/FondOfKudu/Client/CartsRestApi",
-      "FondOfKudu\\Client\\CatalogSearchConnector\\": "bundles/catalog-search-connector/src/FondOfKudu/Client/CatalogSearchConnector",
-      "FondOfKudu\\Shared\\CatalogSearchConnector\\": "bundles/catalog-search-connector/src/FondOfKudu/Shared/CatalogSearchConnector",
-      "FondOfKudu\\Zed\\CatalogSearchConnector\\": "bundles/catalog-search-connector/src/FondOfKudu/Zed/CatalogSearchConnector",
+      "FondOfKudu\\Client\\ProductPageSearchConnector\\": "bundles/product-page-search-connector/src/FondOfKudu/Client/ProductPageSearchConnector",
+      "FondOfKudu\\Shared\\ProductPageSearchConnector\\": "bundles/product-page-search-connector/src/FondOfKudu/Shared/ProductPageSearchConnector",
+      "FondOfKudu\\Zed\\ProductPageSearchConnector\\": "bundles/product-page-search-connector/src/FondOfKudu/Zed/ProductPageSearchConnector",
       "FondOfKudu\\Zed\\CountryZipcodeRestriction\\": "bundles/country-zipcode-restriction/src/FondOfKudu/Zed/CountryZipcodeRestriction",
       "FondOfKudu\\Glue\\DiscountPromotionsRestApi\\": "bundles/discount-promotion-rest-api/src/FondOfKudu/Glue/DiscountPromotionsRestApi",
       "FondOfKudu\\Shared\\DiscountPromotionsRestApi\\": "bundles/discount-promotion-rest-api/src/FondOfKudu/Shared/DiscountPromotionsRestApi",
@@ -151,7 +151,7 @@
       "FondOfKudu\\Zed\\CartsRestApi\\": "bundles/carts-rest-api/tests/FondOfKudu/Zed/CartsRestApi",
       "FondOfKudu\\Zed\\CountryZipcodeRestriction\\": "bundles/country-zipcode-restriction/tests/FondOfKudu/Zed/CountryZipcodeRestriction",
       "FondOfKudu\\Client\\ProductImageStorageConnector\\": "bundles/product-image-storage-connector/tests/FondOfKudu/Client/ProductImageStorageConnector",
-      "FondOfKudu\\Zed\\CatalogSearchConnector\\": "bundles/catalog-search-connector/tests/FondOfKudu/Zed/CatalogSearchConnector",
+      "FondOfKudu\\Zed\\ProductPageSearchConnector\\": "bundles/product-page-search-connector/tests/FondOfKudu/Zed/ProductPageSearchConnector",
       "FondOfKudu\\Zed\\CheckoutRestApi\\": "bundles/checkout-rest-api/tests/FondOfKudu/Zed/CheckoutRestApi",
       "FondOfKudu\\Glue\\CheckoutRestApiCountryConnector\\": "bundles/checkout-rest-api-country-connector/tests/FondOfKudu/Glue/CheckoutRestApiCountryConnector",
       "FondOfKudu\\Zed\\CheckoutRestApiCountryConnector\\": "bundles/checkout-rest-api-country-connector/tests/FondOfKudu/Zed/CheckoutRestApiCountryConnector",
